this works but is not ideal setup
